sample/src/requirements.txt
requests
boto3
loguru
python_ulid